fetch-paged-features is now living in its own file installed eslint, fixed all the things
